Fix the APCU serialization/unserialization handler.
Fixes #21.

And clean up APC-related code - The APC pecl has no php7 releases.
APCU and APCUBC are used instead for those functions and classes.

There may be a leak in the APCU code for custom serializers'
serialization/unserialization methods.
I was getting two clones of an object, instead of one object, for a
while.

Unit tests pass locally with a patched APCU installation,
but fail with the current master branch of APCU.